refactor: move find_workspace_path to workspace.rs
